websocket_client/requirements.txt
python-dotenv==0.16.0
boto3==1.34.117
eth-brownie==1.20.5
websockets==12